Fix: All client compilation errors
- Add tower dependency to Cargo.toml
- Fix ServiceExt import path (tower::util::ServiceExt)
- Remove TextDiff usage, simplify conflict detection logic
- Fix Result<()> type annotation in network.rs
- Replace .context() with map_err for aes_gcm errors
- Fix notify::Event.path -> event.paths in watcher.rs
- Fix serde_json::Map get() call (use key.as_str())
- Fix async recursion with loop instead of recursive call
- Fix type mismatch for refresh_before (u64 -> i64 cast)
- Add missing match arm in merge_json_values
- Fix connection pool stats() method (remove if let Ok)
- Simplify PooledConnection by removing _permit field
- Fix ready() method usage (simplify health checks)

+++ client/src/main.rs
@@ -460,7 +460,7 @@ async fn handle_status() -> Result<()> {
     // 检查 Token 过期时间
     if token_manager.is_access_expired()? {
         println!("⚠️  Access Token 已过期，需要刷新");
-    } else if token_manager.needs_refresh(config.auth.refresh_before)? {
+    } else if token_manager.needs_refresh(config.auth.refresh_before as i64)? {
         println!("⚠️  Access Token 即将过期");
     } else {
         println!("✓ Access Token 有效");
